requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-allauth==0.54.0
django-bootstrap5==22.2
django-crispy-forms==2.0
filelock==3.8.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
oauthlib==3.2.2
Pillow==9.4.0
platformdirs==2.6.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.17.1
Werkzeug==2.2.2
